+++ services/121-service/src/registration/registrations.service.ts
@@ -995,7 +995,6 @@ export class RegistrationsService {
     if (!userId) {
       throw new HttpException('Not authorized.', HttpStatus.UNAUTHORIZED);
     }
-    // const programIds = user.programAssignments.map(p => p.program.id);
 
     const programIds = await this.getProgramIdsUserHasPermission(
       userId,
@@ -1008,8 +1007,13 @@ export class RegistrationsService {
         CustomDataAttributes.whatsappPhoneNumber as string,
       ];
 
-      const phoneNumber =
-        await this.lookupService.lookupAndCorrect(rawPhoneNumber);
+      const phoneNumber = await this.lookupService.lookupAndCorrect(
+        rawPhoneNumber,
+        true,
+      );
+      if (!phoneNumber) {
+        return registrations;
+      }
 
       const matchingRegistrations = (
         await this.registrationRepository.find({