+++ services/121-service/test/helpers/program.helper.ts
@@ -76,7 +76,7 @@ export async function doPayment(
   amount: number,
   referenceIds: string[],
   accessToken: string,
-  filter: { [key: string]: string } = {},
+  filter: Record<string, string> = {},
 ): Promise<request.Response> {
   const queryParams = {};
   if (filter) {
@@ -119,9 +119,9 @@ export async function getTransactions(
   accessToken: string,
 ): Promise<request.Response> {
   return await getServer()
-    .get(`/programs/${programId}/payments/transactions`)
+    .get(`/programs/${programId}/transactions`)
     .set('Cookie', [accessToken])
-    .query({ minPayment: paymentNr, referenceId: referenceId });
+    .query({ payment: paymentNr, referenceId: referenceId });
 }
 
 export async function exportList(
@@ -275,8 +275,8 @@ export async function startCbeValidationProcess(
   accessToken: string,
 ): Promise<request.Response> {
   return await getServer()
-    .get(
-      `/programs/${programId}/financial-service-providers/commercial-bank-ethiopia/account-enquiries/validate`,
+    .post(
+      `/programs/${programId}/financial-service-providers/commercial-bank-ethiopia/account-enquiries/validation`,
     )
     .set('Cookie', [accessToken]);
 }