+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.104.1...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.105.1...master)
+
+---
+
+## [1.105.1](https://github.com/global-121/121-platform/compare/v1.105.0...v1.105.1) - 2023-11-06
+
+### Added
+
+- Filter registrations in PA-table on 'Registration created date'
+
+## [1.105.0](https://github.com/global-121/121-platform/compare/v1.104.2...v1.105.0) - 2023-11-03
+
+### Fixed
+
+- Improved Commercial Bank of Ethiopia error handling
 
 ### Changed
 
@@ -13,13 +27,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Make sure to use your Node version manager to install+run the newest version.
 - Upgraded Node.js to v18 for 121-service.  
   Make sure to recreate the Docker image using `npm run start:services`.
+- Improved performance of 'last payment summary' in payment page
+- Refactored API-paths
+- Prohibit note-creation for referenceIds of programs you do not have permission for
 
 ### Removed
 
 - All audio-features are removed from the PA-App. To prevent a lot of "local changes" during development,
   make sure to remove any previously generated `*.webm`/`*.mp3` audio-files at: `/interfaces/PA-App/src/assets/<language-code>`
 
----
+## [1.104.2](https://github.com/global-121/121-platform/compare/v1.104.1...v1.104.2) - 2023-11-02
+
+### Fixed
+
+- Permission status change fix
 
 ## [1.104.1](https://github.com/global-121/121-platform/compare/v1.104.0...v1.104.1) - 2023-10-31